move modutils to own sub-module
