+++ specparam/modutils/__init__.py
@@ -1,0 +1,1 @@
+"""Module related utilities - generally for internal use."""
